Refactor and fix linting errors.
